Extract and fix query owner of test
