+++ src/lib.rs
@@ -27,11 +27,12 @@ mod unittest_public_metadata;
 mod unittest_queries;
 mod unittest_reveal;
 mod unittest_transfer;
+mod unittest_query_all_tokens;
 mod unittest_query_contract_config;
 mod unittest_query_contract_info;
 mod unittest_query_minters;
-mod unittest_query_all_tokens;
 mod unittest_query_nft_dossier;
+mod unittest_query_owner_of;
 mod unittest_query_num_tokens;
 mod unittest_query_tokens;
 mod unittest_query_unwrapped;
